Point $wgExtensionCredits paths to files instead of directories
This directly follows I2f258ac, which partly solved a problem, and partly
made it worse.

Most (but not all) code in core assumes the "path" points to a regular file,
and strips the last part via dirname(). That's why __DIR__ appears work.
dirname() strips "repo/", "client/", "lib/" and "view/", and what's left is
the Wikibase root directory, where .git/ and CREDITS can be found.

Unfortunately there is also code in core that checks the "path" with is_dir()
first.

https://www.mediawiki.org/wiki/Manual:$wgExtensionCredits suggests to point
to the extension.json file in the root directory, but that doesn't exist here.
Pointing to Wikibase.php instead would be possible, but this is a testing
entry point.

Please also review I9bb5970 which fixes the inconsistent behavior in core.

Bug: T181977
Change-Id: I9f49574c2793d73f6a00f9f9888392ca2c30d5d3