dispatchChanges allow configuration instead of params
This means defaults for dispatchChanges can be controlled
by mediawiki config vars instead of passing in params
to the maint script itslef.

The result of this is within the WMF setup we can run the maint
script with no parameters in operations-puppet and fine tune
the configuration options ourself within mediawiki-config.

Change-Id: If24dd730de6c3bfbf0e47742c447e6da1084db12