Expose JS RevisionStore to custom Changers registered by extension
If the custom changer updates the entity, there might be a need
to update some revision data in the store. In order to allow
changes to have the same revision state, they might use the
same RevisionStore instance. To do so, this makes
EntityChangersFactory expose that instance to whatever user
needs it.

Bug: T181253
Change-Id: Id3899764e6a76889f9344d8d6b79664445d35128