Extract names of search fields as constants
The mapping is currently configured by constructing an array where the
key is the field name. The field itself has a name declared. Extract
the name literals as constants to avoid inconsistencies between the
name declared in the mapping array and the name of the field.
The name of the field is not very important but is used is some
circumtances such as the similarity configuration.

This also fixes an inconsistency for the "descriptions" field which
was exposed as "descriptions" but named "description".

Bug: T182293
Change-Id: I4e79bf1a389beec423b07e34082c4b8920e54af0