Update patch set 6
Patch Set 6:

This patch is 999 lines now. The vast majority of this is tests. One reason for the additional lines of code is that I inlined parts of some overly "clever" helper methods. Now there is a bit more duplication, but the individual test cases are hopefully better readable.

Patch-set: 6