Update patch set 1
Patch Set 1: Code-Review+1

(2 comments)

I'm willing to merge this as it is, but would like to have an other pair of eyes on this. Most preferably by Stas.

The changes to the first parent::__construct() argument look weird at first. Some calls used empty string, one had a typo (as explained in the commit message). The reason this did not broke anything is that this constructor argument is never used anywhere, only in a few tests. I wonder what the original plan for this argument was?

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1