Make wbsetclaim and friends check entity permissions.
wbsetclaim, wbsetclaimvalue, wbsetqualifier, and wbsetreference did not
check user permissions against the actions required by the respective ChangeOp.

Since we currently do not have special permissions defined for modifying Statements,
this was not a problem in practice. BUt it's an incosistency that may lead to
nasty surprises down the road.

Change-Id: Ie0d83165508723f89b3a6e9bac6c4704237c7d5a