Update patch set 3
Patch Set 3:

Looks good to me. 
As a general thought, I think it'd be nice if we could move such configs to mediawiki-config maybe? I imagine these numbers depend a lot on the nature of our data, and it may make sense to have sane defaults (for generic Wikibase users) and highly-tuned config for our own Wikidata.org purposes.
I think I'll make a phab task for it so we could discuss properly.

Patch-set: 3