Update patch set 42
Patch Set 42: Code-Review+1

I think this is ready to go, I have a config patch+reindex that I'd like to deploy but this probably can be deployed without it.
What left to be done is better tuning for language fallbacks as we did for prefix search but this can be done separately imo.

Patch-set: 42
Label: Code-Review=+1