Teach wb.RevisionStore.js to deal with (sub-)entities
This is an alternative for the approach presented in Id389976 and
I7da9e55.

This kind of introduces the concept of sub-entities into the existing
RevisionStore, while avoiding to literally mention the concept (and
having to explain it). Instead, we add a very generic map of "entity
ID => revision number". This is intentionally generic and allows, for
example, to have multiple property views on one page sharing the same
RevisionStore.

Actual use cases this is meant to fulfill:

* When editing statements on forms and senses, the revision IDs for these
statements must be tracked. This works just fine as long as the form/sense
already exists (and because statement GUIDs are globally unique, all
sub-entities can share one RevisionStore). But when a form was just
created, the revision that now contains a form for the first time must be
set via setEntityRevision, and is then used in getClaimRevision.

* In the same scenario, later edits on the same form are not going to
conflict with later edits on statements on that form, because all
statements are either tracked individually via previously called
setClaimRevision, or new (and then it's fine to use any later revision,
as long as it contains the form).

* Edits on top-level elements on a form will be tracked via
setEntityRevision. Such top-level elements are a new concept that never
existed for items and properties

Bug: T181253
Change-Id: Ia603bda92d9e8393e60e64b9e945b790e5e300c5