Update patch set 1
Patch Set 1:

Thanks for the patch,
Have you considered doing this "kind of thing" on WikibaseLexeme's side,  and not in the "core" Wikibase?

Not saying that would be better, it had simply been the idea I had in mind. Would it make any sense if I tried to quickly tinker up the patch so we can have a look?

Patch-set: 1
Label: Code-Review=0