Make (get)HelpMessage optional in $.wikibase.edittoolbar.js
I decided to not make the getter optional, because this allows for a
more strict, clear interface. The getter is guaranteed to be there, it
just might not return anything.

Since calling the getter is derived, I went for hiding the empty question
mark icon after it was created. This should be the exceptional case
anyway. There is usually only 1 of these icons on a page anyway, so no
performance implications because of this.

Bug: T182890
Change-Id: I0c2bebafb163353362d8fb60c5d0c3fb588a4a4b