Update patch set 8
Patch Set 8:

(1 comment)

I added back the part about checking for objects, regarding your patch, I think we should merge this first or rebase it on top of this, as this patch removes lots of usecases in the codebase (specially in prod). Also Would love to hear Marius' reasoning about it.

Patch-set: 8
Label: Verified=0