Update patch set 3
Patch Set 3: Code-Review+1

As far as I can tell this is needed no matter which approach we choose (except for getEntity, which is not needed for every approach). This should be merged along with the base patch that needs it, the moment the decision is made.

Patch-set: 3
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1