Update patch set 5
Patch Set 5:

(1 comment)

. If CI had been allowed to run in between the rebase and the force-merge, this could have been spotted and this change would not have broken CI for unrelated extensions (e.g. https://integration.wikimedia.org/ci/job/mediawiki-extensions-hhvm-jessie/30433/console).

Please be EXTREMELY CAREFUL when force-merging changes in the future, and only do it when absolutely necessary. My understanding is that this was a refactoring commit that was not urgent to merge. In those cases, find out what's breaking CI and fix it, do not force-merge the change, and especially do not blindly force-merge the result of a rebase.

Patch-set: 5