Update patch set 5
Patch Set 5:

Okay, I'm extremely sorry for what happened and will take extra cations next time, FWIW, just one CI was broken and that one was unrelated because master of Wikibase has been failing for a while now and it's keeping us from moving forward (and merging). If I oversight and there was a actually failure in CI, (I just checked and it seems I did), I'm extra sorry for that. Thanks for taking care of the problem

Patch-set: 5