Update patch set 4
Patch Set 4: Code-Review+1

I'm very much fine with this. Amir, do I get it right that *this* test does not leave a dirty database behind? Can you please have a look and tell me if this impression is correct?

This means an other test creates an entry with the title "Dummy2". Isn't it possible to find this test?

Patch-set: 4