Update patch set 6
Patch Set 6:

> I think hard-coding the item and property types is not a good alternative. Configuration allows us to turn something on and of a bit faster, without a backport.

I agree, and as a further comment on that – if I understood it correctly, the idea was to add another change in WikibaseLexeme to edit this config variable during extension initialization. I think it would be much cleaner to leave it at its default value in Wikibase and WikibaseLexeme, and to disable RDF export for lexemes in wmf-config instead.

Patch-set: 6
Reviewer: Lucas Werkmeister (WMDE) <4699@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0