Escape autocomment message arguments
As far as I can tell, none of the existing wikibase-*-summary-* messages
have any wikitext parameters:

- wikibase-entity-summary-wbcreateredirect,
  wikibase-entity-summary-wbmergeitems-from and
  wikibase-entity-summary-wbmergeitems-to only insert the plain page
  title for the mentioned page, not a link.
- wikibase-entity-summary-undo and wikibase-entity-summary-restore
  create a link to Special:Contributions inside the message text itself,
  the parameter is just the plain user name.
- All other message parameters appear to be language codes, wiki codes,
  or counts, all of which should be plain text.

So none of the messages should break when we escape wikitext in the
arguments. On the other hand, some messages take user name arguments,
user names can contain characters like ‘*’ or ‘;’ which at the beginning
of a line have special meaning in wikitext, and the English message
substitutes the user names like {{GENDER:$3|$3}}, which inserts a
newline before the username – with the effect that usernames like “*foo”
or “;bar” render as lists or definition lists in the summary. Escaping
all arguments (in AutoCommentFormatter, we don’t know which arguments
are usernames) fixes this.

(We can’t just use plaintextParams() instead of params(), since that
only substitutes the parameters after parser functions are processed,
i. e. {{COUNT}}, {{GENDER}} etc. are broken.)

Bug: T182800
Change-Id: Iae6bac3bd0d487de9a5ca68e117c070105c695db
Depends-On: Ic5a8fc76c28dca43ce8e334ef1874c2673433f00