Update patch set 1
Patch Set 1: Code-Review+1

Please update the tests. Don't remove them.

I'm fine with this approach and will merge this patch within the next days if nobody objects.

The only minor complain I have is that T182800 is not in our sprint board. Does this mean this is a volunteer patch? Don't we need to track it anyway? Maybe starting from  tomorrows "holiday sprint"?

Patch-set: 1
Label: Code-Review=+1