Update patch set 12
Patch Set 12:

(2 comments)

I'm afraid there is not much I can do here, except for what I now submitted as comments. What I see here looks fine from a distance. However, T182271 is currently not tracked on our sprint board, which means review time I spend on this is not visible to the rest of the team. I also don't know enough about the code base in question, which is not the Wikibase but the CirrusSearch code base. This (again) raises the question if this code should better live in the CirrusSearch extension?

Patch-set: 12