Update patch set 6
Patch Set 6:

So I understand how this would work, and I don't doubt it would be OK.

My point was more like: is disabling RDF output for a particular entity type something which is meant to be the actual "feature", or is it a temporary workaround? If the latter, I am just saying maybe make this workaround simpler and don't bother with config etc. If it is clear at this point there will entity types which are not going to be visible in RDF at all, then the solution like the one here seems like a good way to handle thi.s

I don't know what is the planned or expected behaviour here, that's why I ask.

Patch-set: 6