Update patch set 6
Patch Set 6:

To make things clear for my fellow reviewers: I do not oppose this change. Technically the patch looks complete and ready to go.

I've just asked the question to which I didn't know the answer. The patch neither T182660 do not provide such an answer. My not knowing something is not a reason to not merge the code. If my question is not clear, it is probably a sign it could be easily ignored :)

If others find the solution here good, I believe the patch has got enough of other people's attention, so there will be some one who merges it.

> we possibly need to convince people to merge backports

What do you mean in this context?

Patch-set: 6