Update patch set 6
Patch Set 6:

> I've just asked the question […]

I'm afraid at this point I don't know to which question you are referring to.

> What do you mean in this context?

What I read in your comment above was the suggestion to hard-code the "item" and "property" types, instead of having configuration. This means we can only enable RDF exports for more entity types if we change this hard-coded lines of code. We can do this every week now, which is quite good. But changing configuration can be done even faster. All we need is somebody willing to hit the +2 button on a config change.

If we want to do the same without configuration, we would need to remove the hard-coded restriction, and backport that patch to the branch currently used on production. This is unusual for something that at least feels like a configuration change. We will probably not find anybody willing to merge such a backport. What would be the argument to not wait a week?

Patch-set: 6