Update patch set 6
Patch Set 6:

For transparency: I've just had a little chat with Lydia and asked here the question I refer to, i.e. " is disabling RDF output for a particular entity type something which is meant to be the actual "feature", or is it a temporary workaround?"

What I got from her answer, what is important now is disabling the RDF output for lexeme and co, as it is not complete, buggy, etc, until it is done properly. There might be some other entity types in the future which will be not exposed in RDF as well, but this is very unspecific future, which is not planned at all.

So things are clear to me now. As said above, this patch would solve the task at hand. I believe there might ways to do it even simpler. Which does not make the way introduced here invalid or so.

This was my input to the code review. Thanks for patience.

Patch-set: 6