Update patch set 46
Patch Set 46: Code-Review-2

What I see here is a patch that adds literally threethousand+ lines of CirrusSearch related code to the Wikibase code base, but nobody in the team owning the Wikibase code base was ever really involved in developing this, nor was the ticket ever on any of the Wikidata teams sprint boards. The knowledge about this code is pretty much entirely outside of the team. We just can not maintain this. We don't have the capacity, nor the knowledge, not the processes to be able to balance this load on our shoulders.

The code I see in this patch is glue code between CirrusSearch and Wikibase. A majority is CirrusSearch specific. Sure, it accesses Wikibase concepts, e.g. labels, descriptions, and language fallback chains. But that's pretty much it. Everything else (I would argue it's the majority) is CirrusSearch specific. This and the fact the people and knowledge about this is in the CirrusSearch team, I would argue this belongs to the CirrusSearch code base. Or, even more preferably, to a separate glue code base.

If you want us, the Wikidata team, to integrate this code into the code base we own, maintain, and generally care about, I kindly ask you to be more sensible. We need to be much more integrated into what's going on here. This is going to seriously affect the time and effort we need to invest in the future of this code base.

Here is an other number: This patch is worth >2% of the Wikibase repo code base.

I'm going to set a temporary -2 until we figured this out. This is not meant to block you from continuing to work on this, the contrary. Please do so. I will also happily help with every small refactoring patch we can split from this big patch.

Patch-set: 46