Remove link text normalization from CommonsLinkFormatter
This is critical in diffs (see T99664), but can also be relevant in other
instances that use the same formatter.

This is different for other datatypes like quantities or coordinates. In
these cases normalization happens *before* a value is stored in the database.
The formatters do not do additional normalization. Neither should this one.
I think this makes this behaves more consistent and fits the users
expectations much more.

Bug: T99664
Change-Id: I870271c0b9fa21d2cf7159ca3397f744a0f39644