Update patch set 46
Patch Set 46:

I do not see any way to make it live in CirrusSearch code base. I'd like to hear Lydia 's opinion on this - if we can not work together on improving search on WMDE, then we should not invest so much effort in this only to arrive at the rejection at the end. What's the point of working on the patch if you can not use it?

Patch-set: 46