Update patch set 1
Patch Set 1:

I tested this in localhost and worked just fine, tried to add tests as much as possible but it seems that those parts of frontend code is uncovered (or I couldn't find it).
Regarding exposing the wg variable, It's already exposed in WMF wikis (and VE take advantage of that) but not in all installations (including my localhost) so I explicitly exposed it too.

Patch-set: 1