Update patch set 1
Patch Set 1:

we could rename the setting like this, sure. I believe though the point of T153767 is to integrate "main-repo"-specific settings of client (repoDatabase etc) into the single repositories setting.

I am fine with just renaming now, but it is going to become more confusing like this IMO ("why do I put one database name to repoDatatbase and not to the repositories setting?")

Patch-set: 1