Update patch set 5
Patch Set 5: Code-Review+1

I did the rebase. I also tried to minimize the diff a little bit.

As far as I remember PM was fine with this. The absolute worst case scenario is that three cookies are reset for all anonymous users:
* The license tooltip.
* The "your IP will become public" tooltip.
* If terms are expanded by default.

These cookies are reset anyway after 30 days. That's the main argument why it's not that much of an issue to reset them with a change like this.

Logged in users will not be affected at all.

Patch-set: 5
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1