Properly load collapsible sitelinks CSS exactly when needed
This patch fixes a series of issues:

* The first is the placement of this CSS. It didn't really belong to
"toolbaritem". Note that the relevant "collapsible" class is set on the
"sitelinkgroupview" template.

* This CSS is only needed on JS clients. But it's needed before the
collapsible feature kicks in. Otherwise a brief flash of an unstyled
"[Collapse]" button would be visible. This is achieved by making it it's
own module, and loading it as a dependency.

* The "toolbaritem" is not loaded on diff pages. That's why diff showed
an unstyled "[Collapse]". This is fixed now.

Bug: T175492
Bug: T182767
Change-Id: I6713bdf4c4994475d80c3a7a55cf6550d17b1968