Do not collapse references on diffs/old revisions
This fixes an other regression introduced via Ied5fe1a.

We fixed an other regression just recently, see T181807. Now I realized
we left a lot of code behind that is now entirely unused:
* The "editable" flag is always true now. It can as well be removed.
* Because of this, the EmptyEditSectionGenerator is unused.
* Because of this, the edit section string can not be empty any more.

This patch here just removes all this code.

What's left is the "initially collapsed" state depending on the number
of references. However, this must only have an effect on regular entity
pages that are neither a diff nor an old revision.

The actual fix for T182767 is in the .less file.

Splitting this patch is not impossible, but I feel the proposed solution
can be much better understood with the dead code removed.

Bug: T182767
Change-Id: Ifc8ac520f70914c3ee5e558afc7cfe82b409b737