Update patch set 1
Patch Set 1: Code-Review-1

This one is a weird case and an exception to the rule. Changing the rawParams to plaintextParams is ok, but you need to leave the htmlspecialchars() in there, because the authors want to undo a level of unescaping (See the code comment a couple lines above for details).

So it should be ->plaintextParams( htmlspecialchars( ...

Patch-set: 1
Label: Code-Review=-1