Update patch set 2
Patch Set 2:

Reason was segfaulting on php 5.5 Travis builds when this code is in: https://travis-ci.org/wikimedia/mediawiki-extensions-Wikibase/builds/324099938 or https://travis-ci.org/wikimedia/mediawiki-extensions-Wikibase/jobs/324502806.

Couldn't reproduce this yet, so one possibility is this is actually a glitch on Travis. Revert was made due to unknown cause of the failure. Investigating now.

Patch-set: 2