Update patch set 2
Patch Set 2:

I understand why this was done and as far as I remember I was part of the team which made this decision. However, CSS rules should not be reused for unrelated features like this. This typically turns out to be a bad idea, like it does here now. Please introduce a dedicated line of CSS over there in the constraints extension. Thanks.

Patch-set: 2