Update patch set 1
Patch Set 1: Code-Review-2

Thanks for chiming in Daniel. I agree this does seem very iffy. Actually I was not sure this is a right change to do, and while implementing this I was getting more and more convinced it should not be changed. But I went for leading this to the point it more or less works, and see what others say.

With what you said, I am quite convinced we should not proceed with this change. I am going to abandon this patch soon then.

Patch-set: 1
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-2