Add missing "use" to ChangeNotifier
This is a direct follow up for what I learned in Ie0e0612. Phan reports
this as an error because of the missing "use" clause.

I'm also streamlining this code a bit, e.g. made an unused method private
(we strictly stick to "private by default" in our code bases).

Change-Id: Iafc1d41c7718a90b0df0fbc376be85f6078172ea