Update patch set 1
Patch Set 1:

I see, thanks Ladsgroup.

Would be OK for me to have it like this, at least for start.

To get this clear for me: so the "path" thing is for now only used for diffs containing "subentities" of the lexeme, this path thing was not used in diffs before?

Patch-set: 1