Don't use generic ID parser when a PropertyId is required
This fixes an issue Phan reports.

The generic parser can return any EntityId. But the very next line in
the same method does have a strict type hint against PropertyId.

What this patch effectively does is replacing a possible runtime
exception (an incompatible EntityId can be passed to a method that
expects a PropertyId) with a more expressive parsing exception (an ID
string does not result in a PropertyId).

This code might need more love, but I believe this small step is
already an improvement.

Change-Id: I0066fe188dc53ac9fd3cae3ab523d52477eeec06