Streamline mw.wikibase.entity:getAllStatements
This patch fixes one issue: resolving a label might fail and return nil.
An array access with nil does work just fine, but it's still cleaner and
more obvious if this case is checked before.

This patch also streamlines the code and tries to make it as easy to
follow as possible.

Bug: T166056
Change-Id: Ieed8cedadec4bde0f9a35bac8ebf035ad90c72d0