Update patch set 1
Patch Set 1: Code-Review+1

(2 comments)

I'm not super happy with this proposal for the reasons explained in my inline comments, and because of the position of the new parameter. See, this path is meant to be *prepended*. This is not obvious from the variable name, nor from the position of the parameter. In most methods it should be the *first* parameter, before the actual value, and not be optional.

However, I believe this can be changed in later refactorings.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1