Improve type hints and remove 100% redundant PHPDoc blocks
The main reason for this patch are the missing type hints and & I am
adding in certain places.

I'm also removing PHPDoc block that repeat 100% exactly what the method
header already states. We already did this once, that's why only a few
of such comment blocks could be found for this patch.

Change-Id: I222660937fc7f3ce6a0ea5fbf7198199d6b08bef