Move path array prepending logic to ClaimDifferenceVisualizer
With this patch I am making the prepended path a private implementation
detail of the ClaimDifferenceVisualizer class, and remove it entirely
from DifferencesSnakVisualizer where it was before. That solves all
concerns I had with the original patch Ia4bebd8:

* The $path is not passed so deeply into code that does not really need
  to know anything about it.
* Many of the optional $path parameters are gone.

I had multiple reason to do it this way and would like to explain them,
if needed.

One reason is that a method called "getPropertyHeader" should only do
exactly that: return the substring that describes a property.

Note this patch does not change anything that is needed for I50eb048.
The code I am touching in this patch is not used outside of the
Wikibase code base.

Bug: T182424
Change-Id: I0408672f42b9822f0fe380d8f692b15a7d64f73d