Update patch set 5
Patch Set 5:

I'm not sure if the recursive approach I see here is really needed. Instead, one could traverse the array of usages, and the moment a generic "L" or "D" usage is found, remember that and drop all "L.…" and "D.…" usages for the same entity. This would avoid having *all* aspects twice in memory. Instead, only this information needs to be stored:
* Candidates that might be dropped.
* If a generic "L" or "D" was found.

At the end of the loop, the second array or found generic usages is iterated, and applied on the candidates.

Patch-set: 5