Update patch set 2
Patch Set 2:

re Lucas: Blank nodes are a problem we already have now (with sharding)… just look into the current dump. I wasn't really aware of that… not whether it needs fixing, but I suppose we could, if wanted.

Repeating prefixes etc. is fine, we can do that as often as we want (we already do this once per shard). This is not elegant, but also not a huge problem IMO.

Patch-set: 2