Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

I think this is a good start, and I am willing to merge this.

A fallback to English should also be added, before simply returning the first label. But this can be done in a separate patch.

A test should be added, I believe.

Patch-set: 1
Label: Code-Review=+1