Update patch set 2
Patch Set 2:

That this approach has an OK from Wikidata PM?
Discussion on ticket does not help to judge if this is the way it was meant to be or not.

Technically, the code looks reasonable.

Re testing: there would be several ways to test this behaviour, and testing EntityContent's methods "directly" is only one of these (I am not sure the best one). There definitely should be tests if asking me (does not have to be in this patch if not wanted).

Patch-set: 2