Update patch set 3
Patch Set 3:

There are two references:
* Passing the variable to the array via [ &$foo ] creates a link between the original variable and the array element.
* Accepting a reference in the hook handlers function header creates a link between the array element and the function parameter.

If you want a reference, you need both.

If we don't need a reference, we can freely break it on our side with no consequence. Except there is a tool that knows how MediaWiki hooks works, analyzes them, and complains if the use of & characters does not match. But I'm not aware of such a thing.

Patch-set: 3