Update patch set 3
Patch Set 3: Code-Review+2

I talked to PM in person and got approval for this addition. We need to communicate how to disable this new tracking category in case it is unwanted on a local wiki. This is possible by setting the message to "-", see https://www.mediawiki.org/wiki/Help:Tracking_categories

Patch-set: 3