Update patch set 2
Patch Set 2:

FYI: When adding "minus-x fix" (which this patch doesn't do), Lucas and I found out that the order of the commands should be like this:

"fix": [
	"minus-x fix .",
	"phpcbf"
]

The reason is that phpcbf returns an error code 1 on success (this is already reported upstream, but unlikely to ever be fixed as it would break existing workflows). Under certain circumstances (I think when running this locally), minus-x will not be executed if the previous command returned an error code.

By executing phpcbf last this issue can easily be avoided.

Patch-set: 2