Update patch set 2
Patch Set 2:

The answer to all questions is the same: the *current* code does not modify any of the references I'm touching in this patch, neither unintentionally nor intentional. *Currently* none of these & is needed.

This does not mean they are disallowed to ever be used. If we change one of these hooks in the furure and start doing something with one of these parameters, we will simply add the & back.

As long as this is not the case, the & are not needed.

Removing them does not block anybody from anything.

The only detail you might consider a minor loss is that the method header does not show any more which parameters can be used as references. But documenting these hook handlers never was the job of this code. One must always look back at the docs/hooks.txt file, the relevant MediaWiki.org pages, or even the code triggering the hook to be sure. Relying on the function headers here is calling for trouble.

For all these reasons I don't see a point in splitting this patch, as there was only one reason to remove these &: they are unused here in our code.

Patch-set: 2
Label: Verified=0