Replace mentions of "item" in SetClaim/SetClaimValue tests
These two API modules are meant to work with all entity types that do
have statements. In the tests we are using Items and Properties for
convenience, because these are the default implementations that are both
EntityDocuments and StatementListProviders. But most parts of the test
code don't need to know about this.

Note this patch only touches test code. The two tested classes do not
mention "item" anywhere (except in an example, which is fine).

Change-Id: I2e1843f8cb8a5bb1d5f4833d846717615545468c