Update patch set 1
Patch Set 1:

As the setting is provided by the default file, it is impossible for the new configuration to be missing.
repoDatabase setting will only be there if it was explicitly added to the config (in local settings etc). In such case the assumption is the wiki owner knows what they're doing, so let's use the setting provided, until they migrate their own settings to use the new "repositories" setting.

Will change the commit message in a bit to something less confusing.

Patch-set: 1
Label: Verified=0