Update patch set 3
Patch Set 3:

I don't mind that much. Personally I believe we should keep the number of TestingAccessWrapper usages low. Read: there should be a good reason to test non-public stuff directly. I think this is the case here. A more proper solution would be to make sure the relevant features are all tested via the remaining public methods, and either rework the test cases touched here, or remove them if it turns out they are 100% redundant. But that would be for a later patch.

Feel free to abandon this patch.

Patch-set: 3