Update patch set 2
Patch Set 2:

Because DumpEntities is nothing but an abstract base class, while dumpJson and dumpRdf are actual maintenance scripts. The naming convention for maintenance scripts is to start with a lower case character, while files containing nothing but a class should match the class name and it's capitalization.

Patch-set: 2
Label: Verified=0