Rename Wikibase….hooks.php files to match their class name
"<ExtensionName>.hooks.php" was an old naming scheme many extensions followed.
It became obsolete and is reported as an error by the MediaWiki CodeSniffer
now. The new naming scheme is "<ExtensionName>Hooks.php", identical to the
class name.

I decided to keep the shorter class names and not change "RepoHooks" to
"WikibaseRepoHooks", even if "WikibaseRepo" is the extension name. See
Ib80b820 where the actual extension names are discussed right now.

Change-Id: If86738f66c8df17c795d24e31778cb2554d1c8f7