Remove not-needed & object references from hook handler functions
This is split from Iaaf923e. This patch removes the & from objects.
In current PHP versions it does not make sense to pass objects by
reference. Why would you replace the Parser with an other one, for
example? The reason the & is there is because this was best-practice
in old PHP versions, to avoid expensive cloning of objects.

Change-Id: I41d654681a716ae861107530a7b5a0cacace09d2