Update patch set 2
Patch Set 2:

Adam, if it is really nothing but a typo in a some text, might it be a comment or the commit message, I find it much easier if the reviewer fixes it directly via Gerrits edit feature. In my experience this is the same amount of work as writing and submitting a comment.

The weirdest issue in the specific case here is that a -1 does not go away when I edit the commit message. Therefor I manually removed it for now.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0