Update patch set 2
Patch Set 2:

I think the approach to fix this bug should be to set up a test case that triggers this bug, and then fix it. This way we also have a regression test in place that guarantees the exact same issue can not come back some day on accident. Thanks for considering this.

Patch-set: 2