Fix empty condition list in metadata lookup
WikiPageEntityMetaDataLookup::getWhere() has a special safeguard to make
sure that we don’t accidentally run queries without any WHERE clause.
However, WHERE 0 is not valid SQL syntax (though MySQL and SQLite accept
it), and the MediaWiki Database class interprets a condition of "0" like
an empty condition (no WHERE clause) due to T188314. Instead, use the
condition "0=1", which is valid SQL and guaranteed to be false.

In the test, we trigger this case with an entity ID with no known
associated namespace; without this fix, such a request results in actual
database data for this entity, even though the namespace wasn’t known.

Bug: T188313
Change-Id: I9ad743250f37c3be369888b1b9be80d8d332f62f