Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

The -1 is for the public method name "getPageLatest" which I find not understandable. However, other should feel free to override my -1, especially if the team now taking responsibility for this code base wants to fix this in a later patch instead of amending this one.

Patch-set: 1
Label: Code-Review=-1