Update patch set 1
Abandoned

It makes me sad to see experienced developers not realizing the problem with hard-coded snippets of executable code in configuration strings (a full qualified class name effectively is code). This is guaranteed to break (and already did several times) any time a class is refactored or moved to an other namespace (e.g. to make a codebase PSR-4 compatible). But this is not going to be my problem any more.

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon