[DNM] Move TermBuffer vs. PrefetchingTermLookup violation
To be honest, I do not think what is done in this patch is the best
way to go. What this patch does is moving the violation in one specific
direction. We can also move it in the other direction. Specifically:
I believe all getters returning a "TermBuffer" in the data-access part
of this code base should be changed to return a "PrefetchingTermLookup"
instead. This would instantly fix all violations.

Please have a look at the "TermBuffer" interface. The only getter there
is "getPrefetchedTerm". But this getter is never used outside of
TermBuffer implementations. All usage in production code is via getters
from the PrefetchingTermLookup interface.

In other words: The interface TermBuffer can be ditched and merged into
PrefetchingTermLookup, as this is what the code currently does.

Change-Id: I8a7e7b84454ae0bb878f277561a5aa5b6be5f331