Update patch set 1
Patch Set 1:

@Chad please read discussion in the review, there's serious concern with how this patch is hardcoding the formats and the file extensions and how it is using exceptions as branch condition. 
I completely disagree with the claim that "fully qualified class name is a code" - we use ::class in dozens of places in Mediawiki configs - but even if we wanted to remove this, the patch as it is currently is not a good way to do it. It should be at least seriously amended to fix the problems described.

Patch-set: 1