Update patch set 1
Patch Set 1:

Just to be clear - I am not opposed to changing the config to make it more robust (and would be glad to help but currently don't have much time to drive it) and I see Thiemo's point why. I am just opposed to merging this particular one.

Patch-set: 1